git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/trunk@6814 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74

+++ MagickCore/geometry.c
@@ -1275,8 +1275,8 @@ MagickExport MagickStatusType ParseMetaGeometry(const char *geometry,ssize_t *x,
       (void) ParseGeometry(geometry,&geometry_info);
       area=geometry_info.rho;
       distance=sqrt((double) former_width*former_height);
-      scale.x=former_width/(distance/sqrt((double) area));
-      scale.y=former_height/(distance/sqrt((double) area));
+      scale.x=(double) former_width/(distance/sqrt((double) area));
+      scale.y=(double) former_height/(distance/sqrt((double) area));
       if ((scale.x < (double) *width) || (scale.y < (double) *height))
         {
           *width=(size_t) (former_width/(distance/sqrt(area))+0.5);
